Check self.table is True before trying to get attributes from it
This is for running experiments, which should always have an actionTable: But ImmediateModeExperiments do not have an actionTable, so there needs to be a way to not have an actionTable defined for the experiment to run.
Perhaps a better way would be to make sure self.createValidActionTable() at least returns an empty table?