doc: merge the two very similar experiment examples (#720)
The two immediateModeExample files are very similar.  This commit
merges the two into a single example file.  It's not that any of them
is correct, both were wrong, and it's not that now the merged file is
correct.  It's only that now we only have a single example file to
correct.